Make parallel assembly work
